Add new plot directory to handle actually plotting; reference SvgPlot in home.rs
