+++ trope-web/src/plot/mod.rs
@@ -1,0 +1,3 @@
+mod svg_plot;
+
+pub use svg_plot::SvgPlot;
